+++ request.go
@@ -10,7 +10,7 @@ import (
 	"golang.org/x/net/publicsuffix"
 )
 
-const Version = "0.2.0"
+const Version = "0.3.0"
 
 type FileField struct {
 	FieldName string
@@ -35,6 +35,10 @@ type Args struct {
 	BasicAuth BasicAuth
 }
 
+type Request struct {
+	*Args
+}
+
 func NewArgs(c *http.Client) *Args {
 	if c.Jar == nil {
 		options := cookiejar.Options{
@@ -55,6 +59,10 @@ func NewArgs(c *http.Client) *Args {
 		Proxy:     "",
 		BasicAuth: BasicAuth{},
 	}
+}
+
+func NewRequest(c *http.Client) *Request {
+	return &Request{NewArgs(c)}
 }
 
 func newURL(u string, params map[string]string) string {
@@ -118,11 +126,23 @@ func Get(url string, a *Args) (resp *Response, err error) {
 	return
 }
 
+// url can be string or *url.URL or ur.URL
+func (req *Request) Get(url interface{}) (resp *Response, err error) {
+	resp, err = Get(url2string(url), req2arg(req))
+	return
+}
+
 // Head issues a HEAD to the specified URL.
 //
 // Caller should close resp.Body when done reading from it.
 func Head(url string, a *Args) (resp *Response, err error) {
 	resp, err = newRequest("HEAD", url, a)
+	return
+}
+
+// url can be string or *url.URL or ur.URL
+func (req *Request) Head(url interface{}) (resp *Response, err error) {
+	resp, err = Head(url2string(url), req2arg(req))
 	return
 }
 
@@ -134,11 +154,23 @@ func Post(url string, a *Args) (resp *Response, err error) {
 	return
 }
 
+// url can be string or *url.URL or ur.URL
+func (req *Request) Post(url interface{}) (resp *Response, err error) {
+	resp, err = Post(url2string(url), req2arg(req))
+	return
+}
+
 // Put issues a PUT to the specified URL.
 //
 // Caller should close resp.Body when done reading from it.
 func Put(url string, a *Args) (resp *Response, err error) {
 	resp, err = newRequest("PUT", url, a)
+	return
+}
+
+// url can be string or *url.URL or ur.URL
+func (req *Request) Put(url interface{}) (resp *Response, err error) {
+	resp, err = Put(url2string(url), req2arg(req))
 	return
 }
 
@@ -150,11 +182,23 @@ func Patch(url string, a *Args) (resp *Response, err error) {
 	return
 }
 
+// url can be string or *url.URL or ur.URL
+func (req *Request) Patch(url interface{}) (resp *Response, err error) {
+	resp, err = Patch(url2string(url), req2arg(req))
+	return
+}
+
 // Delete issues a DELETE to the specified URL.
 //
 // Caller should close resp.Body when done reading from it.
 func Delete(url string, a *Args) (resp *Response, err error) {
 	resp, err = newRequest("DELETE", url, a)
+	return
+}
+
+// url can be string or *url.URL or ur.URL
+func (req *Request) Delete(url interface{}) (resp *Response, err error) {
+	resp, err = Delete(url2string(url), req2arg(req))
 	return
 }
 
@@ -164,4 +208,38 @@ func Delete(url string, a *Args) (resp *Response, err error) {
 func Options(url string, a *Args) (resp *Response, err error) {
 	resp, err = newRequest("OPTIONS", url, a)
 	return
+}
+
+// url can be string or *url.URL or ur.URL
+func (req *Request) Options(url interface{}) (resp *Response, err error) {
+	resp, err = Options(url2string(url), req2arg(req))
+	return
+}
+
+func url2string(u interface{}) string {
+	switch u.(type) {
+	case string:
+		return u.(string)
+	case url.URL:
+		s := u.(url.URL)
+		return s.String()
+	case *url.URL:
+		s := u.(*url.URL)
+		return s.String()
+	}
+	return ""
+}
+
+func req2arg(req *Request) (a *Args) {
+	return &Args{
+		Client:    req.Client,
+		Headers:   req.Headers,
+		Cookies:   req.Cookies,
+		Data:      req.Data,
+		Params:    req.Params,
+		Files:     req.Files,
+		Json:      req.Json,
+		Proxy:     req.Proxy,
+		BasicAuth: req.BasicAuth,
+	}
 }